at86rf230: remove unnecessary / dead code
In at86rf230_probe() lp was first set to dev->priv and a few lines later
dev->priv was set to lp again, without changing lp in between. The call
to ieee802154_unregister_device() before err_irq: was unreachable. The
last call which could fail before this line is the call to
ieee802154_register_device().

Signed-off-by: Sascha Herrmann <sascha@ps.nvbi.de>