at86rf230: changed irq handling to not disable irq for edge type interrupt
If interrupt is configured to edge type, do not disable irq in the isr.
The at86rf230 resets the irq line only after the irq status register is
read. Disabling the irq can lock the driver in situations where a irq is
set by the radio while the driver is still reading the frame buffer.
Additional the irq filter register is set to filter out all unused
interrupts and the irq status register is read in the probe function to
clear the irq line.

Signed-off-by: Sascha Herrmann <sascha@ps.nvbi.de>